fix build error (signed/unsigned)

+++ onnxruntime/core/mlas/lib/intrinsics/avx2/qladd_avx2.cpp
@@ -85,9 +85,9 @@ MlasLoad32Bytes(const uint8_t* buffer, int64_t N)
     } else {
         uint8_t dup[32];
         uint8_t* dst = dup;
-        while (N >= sizeof(uint32_t)) {
+        while (N >= (int64_t)sizeof(uint32_t)) {
             *(uint32_t*)(dst) = *(uint32_t*)(buffer);
-            N -= sizeof(uint32_t);
+            N -= (int64_t)sizeof(uint32_t);
             buffer += sizeof(uint32_t);
             dst += sizeof(uint32_t);
         }
